+++ src/core/hiwire.c
@@ -231,10 +231,6 @@ EM_JS_REF(JsRef, hiwire_double, (double val), {
   return Hiwire.new_stack(val);
 });
 
-EM_JS_REF(JsRef, hiwire_string_utf8, (const char* ptr), {
-  return Hiwire.new_stack(UTF8ToString(ptr));
-});
-
 EM_JS(void _Py_NO_RETURN, hiwire_throw_error, (JsRef iderr), {
   throw Hiwire.pop_value(iderr);
 });
@@ -293,6 +289,32 @@ EM_JS_REF(JsRef,
   return Hiwire.new_value(func.apply(this_, args));
 });
 // clang-format on
+
+// Either syncifyHandler will get filled in by stack_switching/suspenders.mjs or
+// stack switching is not available so syncify will always return an error in
+// JsProxy.c and syncifyHandler will never be called.
+EMSCRIPTEN_KEEPALIVE JsRef (*syncifyHandler)(JsRef idpromise) = NULL;
+
+EM_JS(void, hiwire_syncify_handle_error, (void), {
+  if (!Module.syncify_error) {
+    // In this case we tried to syncify in a context where there is no
+    // suspender. JsProxy.c checks for this case and sets the error flag
+    // appropriately.
+    return;
+  }
+  Module.handle_js_error(Module.syncify_error);
+  delete Module.syncify_error;
+})
+
+JsRef
+hiwire_syncify(JsRef idpromise)
+{
+  JsRef result = syncifyHandler(idpromise);
+  if (result == 0) {
+    hiwire_syncify_handle_error();
+  }
+  return result;
+}
 
 EM_JS_BOOL(bool, hiwire_HasMethod, (JsRef obj_id, JsRef name), {
   // clang-format off
@@ -512,12 +534,6 @@ EM_JS_BOOL(bool, hiwire_get_bool, (JsRef idobj), {
     return false;
   }
   return true;
-  // clang-format on
-});
-
-EM_JS_BOOL(bool, hiwire_is_function, (JsRef idobj), {
-  // clang-format off
-  return typeof Hiwire.get_value(idobj) === 'function';
   // clang-format on
 });
 
@@ -818,103 +834,6 @@ EM_JS_REF(JsRef, JsObject_New, (), {
   return Hiwire.new_value({});
 });
 // clang-format on
-
-void
-setReservedError(char* action, char* word)
-{
-  PyErr_Format(PyExc_AttributeError,
-               "The string '%s' is a Python reserved word. To %s an attribute "
-               "on a JS object called '%s' use '%s_'.",
-               word,
-               action,
-               word,
-               word);
-}
-
-EM_JS(bool, isReservedWord, (int word), {
-  if (!Module.pythonReservedWords) {
-    Module.pythonReservedWords = new Set([
-      "False",  "await", "else",     "import", "pass",   "None",    "break",
-      "except", "in",    "raise",    "True",   "class",  "finally", "is",
-      "return", "and",   "continue", "for",    "lambda", "try",     "as",
-      "def",    "from",  "nonlocal", "while",  "assert", "del",     "global",
-      "not",    "with",  "async",    "elif",   "if",     "or",      "yield",
-    ])
-  }
-  return Module.pythonReservedWords.has(word);
-})
-
-/**
- * action: a javascript string, one of get, set, or delete. For error reporting.
- * word: a javascript string, the property being accessed
- */
-EM_JS(int, normalizeReservedWords, (int word), {
-  // clang-format off
-  // 1. if word is not a reserved word followed by 0 or more underscores, return
-  //    it unchanged.
-  const noTrailing_ = word.replace(/_*$/, "");
-  if (!isReservedWord(noTrailing_)) {
-    return word;
-  }
-  // 2. If there is at least one trailing underscore, return the word with a
-  //    single underscore removed.
-  if (noTrailing_ !== word) {
-    return word.slice(0, -1);
-  }
-  // 3. If the word is exactly a reserved word, return it unchanged
-  return word;
-  // clang-format on
-});
-
-EM_JS_REF(JsRef, JsObject_GetString, (JsRef idobj, const char* ptrkey), {
-  const jsobj = Hiwire.get_value(idobj);
-  const jskey = normalizeReservedWords(UTF8ToString(ptrkey));
-  const result = jsobj[jskey];
-  // clang-format off
-  if (result === undefined && !(jskey in jsobj)) {
-    // clang-format on
-    return ERROR_REF;
-  }
-  return Hiwire.new_value(result);
-});
-
-// clang-format off
-EM_JS_NUM(errcode,
-JsObject_SetString,
-(JsRef idobj, const char* ptrkey, JsRef idval),
-{
-  let jsobj = Hiwire.get_value(idobj);
-  let jskey = normalizeReservedWords(UTF8ToString(ptrkey));
-  let jsval = Hiwire.get_value(idval);
-  jsobj[jskey] = jsval;
-});
-// clang-format on
-
-EM_JS_NUM(errcode, JsObject_DeleteString, (JsRef idobj, const char* ptrkey), {
-  let jsobj = Hiwire.get_value(idobj);
-  let jskey = normalizeReservedWords(UTF8ToString(ptrkey));
-  delete jsobj[jskey];
-});
-
-EM_JS_REF(JsRef, JsObject_Dir, (JsRef idobj), {
-  let jsobj = Hiwire.get_value(idobj);
-  let result = [];
-  do {
-    // clang-format off
-    const names = Object.getOwnPropertyNames(jsobj);
-    result.push(...names.filter(
-      s => {
-        let c = s.charCodeAt(0);
-        return c < 48 || c > 57; /* Filter out integer array indices */
-      }
-    )
-    // If the word is a reserved word followed by 0 or more underscores, add an
-    // extra underscore to reverse the transformation applied by normalizeReservedWords.
-    .map(word => isReservedWord(word.replace(/_*$/, "")) ? word + "_" : word));
-    // clang-format on
-  } while (jsobj = Object.getPrototypeOf(jsobj));
-  return Hiwire.new_value(result);
-});
 
 EM_JS_REF(JsRef, JsObject_Entries, (JsRef idobj), {
   let jsobj = Hiwire.get_value(idobj);